df plugin: Makes "/" mountpoint plugin instance configurable
The plugin chooses the string "root" as the plugin instance of the "/"
("primary hierarchy root" and "root directory" in FHS structure).

This causes conflict when a "/root" mountpoint exists (which also gets
"root" as plugin instance).

This commits makes this "root" string configurable so that it can be
changed in case of conflicting scenario

Issue: #739